+++ src/main/java/com/web/happyhouse/house/contorller/AptOfficeController.java
@@ -4,9 +4,6 @@ import com.web.happyhouse.house.domain.HouseType;
 import com.web.happyhouse.house.dto.response.HouseInfoListRs;
 import com.web.happyhouse.house.dto.response.HouseOnSaleDetailRs;
 import com.web.happyhouse.house.dto.response.HouseOnSaleListRs;
-import com.web.happyhouse.house.dto.createRequest.HouseOnSaleCreateRq;
-import com.web.happyhouse.house.dto.response.HouseOnSaleRs;
-import com.web.happyhouse.house.dto.updateRequest.HouseOnSaleUpdateRq;
 import com.web.happyhouse.house.service.AptOfficeService;
 import com.web.happyhouse.network.ResponseCode;
 import com.web.happyhouse.network.ResponseDto;
@@ -14,8 +11,6 @@ import io.swagger.annotations.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
-
-import javax.validation.Valid;
 
 @Api(tags ={"아파트/오피스텔 Controller"})
 @RestController
@@ -68,56 +63,6 @@ public class AptOfficeController {
     })
     public ResponseDto<HouseOnSaleDetailRs> getOnSaleDetail(@PathVariable("houseOnSaleId") Long houseOnSaleId){
         return ResponseDto.res(ResponseCode.OK, aptOfficeService.getHouseOnSaleDetail(houseOnSaleId));
-    }
-
-    // TODO: 부동산중개인 권한설정 필요
-    @ApiOperation(value = "오피스텔/아파트 매물 찾기", notes="오피스텔/아파트 매물 찾기")
-    @GetMapping("/mng/{houseOnSaleId}")
-    @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "OK 오피스텔/아파트 매물찾기를 완료했습니다."),
-            @ApiResponse(code = 400, message = "BAD_REQUEST 선택한 아파트/오피스텔 매물찾기를 실패했습니다.")
-    })
-    @ResponseStatus(HttpStatus.OK)
-    public ResponseDto<HouseOnSaleRs> getHouseOnSale(@PathVariable("houseOnSaleId") Long houseOnSaleId){
-        return ResponseDto.res(ResponseCode.OK, "선택한 아파트/오피스텔 매물찾기를 완료했습니다.", aptOfficeService.getHouseOnSale(houseOnSaleId));
-    }
-
-    @ApiOperation(value = "오피스텔/아파트 매물 생성", notes="오피스텔/아파트 매물 생성")
-    @PostMapping("/mng")
-    @ApiResponses(value = {
-            @ApiResponse(code = 201, message = "CREATED 오피스텔/아파트 매물생성을 완료했습니다."),
-            @ApiResponse(code = 400, message = "BAD_REQUEST 선택한 아파트/오피스텔정보 찾기를 실패했습니다.")
-    })
-    @ResponseStatus(HttpStatus.CREATED)
-    public ResponseDto<Long> createHouseOnSale(
-            @ApiParam(value = "생성 요청 form", required = true)
-            @Valid @RequestBody HouseOnSaleCreateRq createRq){
-        return ResponseDto.res(ResponseCode.CREATED, "선택한 오피스텔/아파트 매물생성을 완료했습니다.", aptOfficeService.saveHouseOnSale(createRq));
-    }
-
-    @ApiOperation(value = "오피스텔/아파트 매물 업데이트", notes="오피스텔/아파트 매물 업데이트")
-    @PatchMapping("/mng")
-    @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "OK 오피스텔/아파트 매물업데이트를 완료했습니다."),
-            @ApiResponse(code = 400, message = "BAD_REQUEST 선택한 아파트/오피스텔 매물찾기를 실패했습니다.")
-    })
-    @ResponseStatus(HttpStatus.OK)
-    public ResponseDto<Long> updateHouseOnSale(
-            @ApiParam(value = "업데이트 요청 form", required = true)
-            @Valid @RequestBody HouseOnSaleUpdateRq updateRq){
-        return ResponseDto.res(ResponseCode.OK, "선택한 오피스텔/아파트 매물업데이트를 완료했습니다.", aptOfficeService.updateHouseOnSale(updateRq));
-    }
-
-    @ApiOperation(value = "오피스텔/아파트 매물 삭제", notes="오피스텔/아파트 매물 삭제")
-    @DeleteMapping("/mng/{houseOnSaleId}")
-    @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "OK 오피스텔/아파트 매물삭제를 완료했습니다."),
-            @ApiResponse(code = 400, message = "BAD_REQUEST 선택한 아파트/오피스텔 매물찾기를 실패했습니다.")
-    })
-    @ResponseStatus(HttpStatus.OK)
-    public ResponseDto deleteHouseOnSale(@PathVariable("houseOnSaleId") Long houseOnSaleId){
-        aptOfficeService.deleteHouseOnSale(houseOnSaleId);
-        return ResponseDto.res(ResponseCode.OK, "선택한 오피스텔/아파트 매물삭제를 완료했습니다.");
     }
 
 }
